New Demo that demonstrates remembered objects.
Latest commit:
Here is a working version of the SceneManager.  It will change between
"Level One" and "Level Two" by pressing the keys 1 and 2.  It remembers
the Scientist
object's position but it currently carries over the direction the Sprite
is facing when changing levels (in "Level One"" if you face left, then
switch to
"Level Two" and face right, if you switch back to "Level One" it will
now be facing right).  I do not think this will be hard to fix, but I
have other
things to work on at the moment.