Eric pointed out a bug, should be fixed.
Eric just pointed out an important bug I had in the Background class. I
had #include "res_path.h" in both the RunGame.exe and the Background
class. This
caused linker errors.  I modified the Background class so that it
doesn't include res_path.h so now when constructing a Background object
you have
to give it the full path instead of just the name of the file.
Special Thanks to Eric for the catch.